requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
backoff
bcrypt
beautifulsoup4
black
cachetools
certifi
cffi
charset-normalizer
chess
chromadb
click
colorama
coloredlogs
config
coverage
cryptography
dataclasses-json
distro
dnspython
durationpy
ecdsa
email-validator
exceptiongroup
faiss-cpu
fastapi
filelock
flake8
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
groq
grpcio
h11
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
iniconfig
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
langchain-text-splitters
langsmith
lxml
markdown-it-py
marshmallow
mccabe
mdurl
mmh3
mpmath
multidict
mypy
mypy_extensions
numpy
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
passlib
pathspec
platformdirs
pluggy
posthog
prometheus_client
propcache
protobuf
psutil
pyasn1
pyasn1_modules
pybase64
pycodestyle
pycparser
pydantic
pydantic-settings
pydantic_core
pyflakes
Pygments
PyPDF2
PyPika
pyproject_hooks
pyreadline3
pytest
pytest-asyncio
pytest-cov
python-dateutil
python-docx
python-dotenv
python-jose
python-multipart
PyYAML
redis
celery>=5.3.0
kombu>=5.3.0
flower>=2.0.0
psycopg2-binary>=2.9.0
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
sarvamai

# --- Audio Providers ---
# Deepgram SDK for real-time STT
deepgram-sdk>=3.0.0
# ElevenLabs for high-quality TTS
elevenlabs>=1.0.0

shellingham
six
sniffio
soupsieve
SQLAlchemy
starlette
structlog
sympy
tenacity
tiktoken
tokenizers
tomli
tqdm
typer
typing-inspect
typing-inspection
typing_extensions
urllib3
uvicorn==0.23.2
chromadb>=0.5.0
chromadb-client>=0.5.0
watchfiles
websocket-client
websockets
yarl
zipp
zstandard
